Resolve TODO
After both inspecting the source and testing, it appears that what
I hoped would be the case (and originally documented) is indeed the
case. In robotd we call `str` on each argument before sending them
over serial, so numbers (etc.) are converted to strings.

The alternative I'd wondered about was integers arriving a raw
integer bytes, rather than as their ASCII representation.